Add function to download pdf and download based the downloadBtn parameter
A new taxonomy for the download button should be added in the base configuration file.
eg downloadable = "downloadBtn"
When this downloadBtn is added in the front matter of any of the content files, it will be displayed.
eg downloadBtn: "true"
Adding it to the archetypes and setting it to false by default be awesome.

Signed-off-by: Ida Delphine <mida@unicef.org>