:memo: layouts/index: Add optional text above site topics
This commit allows a site to have generic text to appear at the top of
the site, underneath the header but above the topics. This is useful
because it allows a site maintainer to give more context or lead-ins to
the content available on the site. Since this might not be applicable
for all users, this is made an optional feature. If there is no index
file in `content/`, nothing will appear and the site will appear as it
does now.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>